requirements.txt
langchain>=0.3
langchain-anthropic>=0.3
langchain-core>=0.3
langgraph>=0.2
anthropic>=0.40
tavily-python>=0.5
yfinance>=0.2
python-dotenv>=1.0
